Extract workspace automation model API
